+++ presentation/src/main/java/co/orange/presentation/setting/delivery/DeliveryActivity.kt
@@ -1,15 +1,13 @@
 package co.orange.presentation.setting.delivery
 
+import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import co.orange.core.base.BaseActivity
 import co.orange.core.extension.breakLines
 import co.orange.core.extension.setOnSingleClickListener
-import co.orange.core.extension.toast
 import co.orange.domain.entity.response.AddressInfoModel
-import co.orange.presentation.address.AddressWebActivity
-import co.orange.presentation.address.AddressWebBridge.Companion.ADDRESS
-import co.orange.presentation.address.AddressWebBridge.Companion.ZIPCODE
+import co.orange.presentation.address.AddressActivity
 import kr.genti.presentation.R
 import kr.genti.presentation.databinding.ActivityDeliveryBinding
 
@@ -20,36 +18,26 @@ class DeliveryActivity : BaseActivity<ActivityDeliveryBinding>(R.layout.activity
         super.onCreate(savedInstanceState)
 
         initBackBtnListener()
-        initAddBtnListener()
-        initModBtnListener()
+        initWebBtnListener()
         initDeleteBtnListener()
         setDeliveryUi(viewModel.mockAddressModel)
-    }
-
-    override fun onStart() {
-        super.onStart()
-        AddressWebActivity.register(this)
     }
 
     private fun initBackBtnListener() {
         binding.btnBack.setOnSingleClickListener { finish() }
     }
 
-    private fun initAddBtnListener() {
-        binding.btnAddDelivery.setOnSingleClickListener {
-            AddressWebActivity.open { bundle ->
-                val zipCode = bundle.getString(ZIPCODE)
-                val address = bundle.getString(ADDRESS)
-                if (zipCode?.isNotEmpty() == true && address?.isNotEmpty() == true) {
-                    toast("주소: [$zipCode] $address")
-                }
-            }
+    private fun initWebBtnListener() {
+        with(binding) {
+            btnDeliveryAdd.setOnSingleClickListener { navigateToAddressView() }
+            btnDeliveryMod.setOnSingleClickListener { navigateToAddressView() }
         }
     }
 
-    private fun initModBtnListener() {
-        // TODO
-        binding.btnDeliveryMod.setOnSingleClickListener { }
+    private fun navigateToAddressView() {
+        Intent(this, AddressActivity::class.java).apply {
+            startActivity(this)
+        }
     }
 
     private fun initDeleteBtnListener() {
@@ -68,10 +56,5 @@ class DeliveryActivity : BaseActivity<ActivityDeliveryBinding>(R.layout.activity
                 ).breakLines()
             tvDeliveryPhone.text = item.phone
         }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        AddressWebActivity.unregister()
     }
 }
